--- lib/utils/baseurl.ts
@@ -1,2 +1,0 @@
-export const baseUrl =
-  process.env.VERCEL_URL ?? process.env.NEXT_PUBLIC_API_URL;
